Move ebulletin logic into own controller
